// src/screens/applications/applicationstageoverviewdetails/application.d.ts
export interface TableRow {
    job_name: string;
    total_applicants: number;
    close_date: string;
    is_closed: boolean;
  
    stages: {
      resume_screening: number;
      assessment_test: number;
      video_interview: number;
      hired: number;
      reject: number;
      on_hold: number;
    };
  }